LayoutTests/css3/calc/simplification-expected.txt

This tests parse time simplification in calc()

100px * (25 + 5) => calc(3000px)
100em * (25 - 5) => calc(2000em)
100ex * (2 * 5 - 5) => calc(500ex)
100cm * (5 - 4 / 5) => calc(15874.015748031496px)
100mm * (2.4 * 5 - 8 / 5) => calc(3930.708661417323px)
100in * (6 * (5 - 4) / 8) => calc(7200px)
1px * (3 + 1/(7 + 1/(15 + 1/(1 + 1/(292 + 1/(1 + 1/(1 + 1/(1 + 1)))))))) => calc(3.1415926535583574px)
100pc * 20 + 100rem * 10 - 100ch * 5 + 100pc => calc(-500ch + 33600px + 1000rem)
((100px + 20 * 5px) * 10 - 5 * (10em * 5 + 10em)) * 2 => calc(2 * (-300em + 2000px))
100px + 1in => calc(196px)
10 * 10px + 0.5 * 2in => calc(196px)
100px + 1in + 10% => calc(10% + 196px)
100px - 1in => calc(4px)
50cm + 50cm => calc(3779.5275590551178px)
50cm + 10in + 100mm => calc(3227.7165354330705px)
100px + 1em => calc(1em + 100px)
100px + 1em + 100px => calc(1em + 200px)
1em + 1rem => calc(1em + 1rem)
